+++ src/config/logger.js
@@ -8,7 +8,7 @@ module.exports = function() {
 
 	self.logger = require('tracer').colorConsole(
         {	
-        	methods : ['log','debug','warn','error','load','test','save','remove','bank'],
+        	methods : ['log','debug','warn','error','load','test','save','remove','bank','io'],
             format : [
                   colors.white("{{timestamp}} ")+colors.yellow(self.botName)+": {{message}}",
                   {
@@ -18,7 +18,8 @@ module.exports = function() {
                       test : colors.white("{{timestamp}} ")+colors.bgRed(self.botName)+": "+colors.white("{{message}}"),
                       save : colors.white("{{timestamp}} ")+colors.green(self.botName)+": "+colors.white("{{message}}"),
                       remove : colors.white("{{timestamp}} ")+colors.red(self.botName)+": "+colors.white("{{message}}"),
-                      bank : colors.white("{{timestamp}} ")+colors.blue(self.botName)+": "+colors.white("{{message}}")
+                      bank : colors.white("{{timestamp}} ")+colors.blue(self.botName)+": "+colors.white("{{message}}"),
+                      io : colors.white("{{timestamp}} ")+colors.blue(self.botName)+": "+colors.white("{{message}}")                  
                   }
             ],
             filters : {
@@ -27,23 +28,24 @@ module.exports = function() {
 	            info : colors.green,
 	            warn : [ colors.yellow, colors.bold ],
 	            error : [ colors.red, colors.bold ],
-	            test : colors.yellow
+	            test : colors.yellow,
+              io : colors.bold
 	        },
 	        level : 'log',
-            dateformat : "HH:MM:ss.L",
-            preprocess :  function(data) {
-              data.timestamp = moment(new Date()).format('MM/DD/YYYY-')+data.timestamp;
-              data.title = data.title.toLowerCase();
+          dateformat : "HH:MM:ss.L",
+          preprocess :  function(data) {
+            data.timestamp = moment(new Date()).format('MM/DD/YYYY-')+data.timestamp;
+            data.title = data.title.toLowerCase();
+          },
+          transport : [
+            function (data) {
+            	// Console
+              console.log(data.output);
             },
-            transport : [
-                function (data) {
-                	// Console
-                  console.log(data.output);
-                },
-                function (data) {
-                  // Static
-                  fs.appendFileSync(self.logs_file, data.output + '\n');
-                },
-            ]
+            function (data) {
+              // Static
+              fs.appendFileSync(self.logs_file, data.output + '\n');
+            }
+          ]
         });
 }